requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
Django==4.2
django-admin-interface==0.24.2
django-colorfield==0.8.0
django-filter==23.2
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==9.5.0
psycopg2==2.9.6
sqlparse==0.4.3
typing_extensions==4.5.0
whitenoise==6.4.0
